serial_controller/requirements.txt
pyserial==3.4
aiohttp==3.7.3
paho-mqtt==1.5.1
#PyYAML==5.1
#websockets==7.0